feat: parallel resolution (#338)
Stencil spends a little while just doing network requests to gather its
dependencies.

Anecdotally, this is like several seconds per module, which makes
resolution take a long time.

This change set adds a smidge of parallelism to that resolution. The
strategy is to launch a fixed number of worker routines, use them to
resolve the work list, and then on completion, check whether items were
added to the work list that no worker picked up, due to a race condition
between when workers check for more work and when work can get added.

Existing tests covered this changeset completely, including revealing
situations where we'd leave items in the work set, so I didn't feel the
need to add additional coverage.